storage leaf generalized for short and long RLPs
